+++ src/main/java/io/takamaka/messages/utils/ChatUtils.java
@@ -124,69 +124,6 @@ public class ChatUtils {
                 .readValue(jsonMessage, SignedMessageBean.class);
     }
 
-    public static final SignedMessageBean verifySignedMessage(String messageJson, String... from) throws ChatMessageException {
-        try {
-            SignedMessageBean fromJsonToSignedMessageBean = ChatUtils.fromJsonToSignedMessageBean(messageJson);
-            final String jsonCanonical;
-            final String pk;
-            switch (from.length) {
-                case 0:
-                    pk = fromJsonToSignedMessageBean.getFrom();
-                    break;
-                case 1:
-                    pk = from[0];
-                    break;
-                default:
-                    throw new ChatMessageException("invalid parameters number, expected 0..1 got " + Arrays.toString(from));
-            }
-            final TkmCypherBean verifyResult;
-            final SignedMessageBean returnObj;
-            switch (fromJsonToSignedMessageBean.getMessageType()) {
-                //java 17 limitation...
-                case "REGISTER_USER_SIGNED_REQUEST":
-                    RegisterUserRequestBean fromJsonToRegisterUserRequestBean = ChatUtils.fromJsonToRegisterUserRequestBean(messageJson);
-                    jsonCanonical = SimpleRequestHelper.getCanonicalJson(fromJsonToRegisterUserRequestBean.getRegisterUserRequestSignedContentBean());
-                    returnObj = fromJsonToRegisterUserRequestBean;
-                    break;
-                case "REQUEST_USER_KEYS":
-                    RequestUserKeyRequestBean fromJsonToRequestUserKeyRequestBean = ChatUtils.fromJsonToRequestUserKeyRequestBean(messageJson);
-                    jsonCanonical = SimpleRequestHelper.getCanonicalJson(fromJsonToRequestUserKeyRequestBean.getRequestUserKeyRequestBeanSignedContent());
-                    returnObj = fromJsonToRequestUserKeyRequestBean;
-                    break;
-                case "TOPIC_CREATION":
-                    CreateConversationRequestBean fromJsonToCreateConversationRequest = ChatUtils.fromJsonToCreateConversationRequest(messageJson);
-                    jsonCanonical = SimpleRequestHelper.getCanonicalJson(fromJsonToCreateConversationRequest.getTopic());
-                    returnObj = fromJsonToCreateConversationRequest;
-                    break;
-                case "TOPIC_MESSAGE":
-                    BasicMessageRequestBean fromJsonToBasicMessageBeanRequest = ChatUtils.fromJsonToBasicMessageBeanRequest(messageJson);
-                    jsonCanonical = SimpleRequestHelper.getCanonicalJson(fromJsonToBasicMessageBeanRequest.getBasicMessageSignedContentBean());
-                    returnObj = fromJsonToBasicMessageBeanRequest;
-                    break;
-
-                default:
-                    throw new UnsupportedChatMessageTypeException("unsupported message type" + fromJsonToSignedMessageBean.getMessageType());
-            }
-
-            switch (fromJsonToSignedMessageBean.getSignatureType()) {
-                case "Ed25519BC":
-                    verifyResult = TkmCypherProviderBCED25519.verify(pk, fromJsonToSignedMessageBean.getSignature(), jsonCanonical);
-                    break;
-
-                default:
-                    throw new UnsupportedSignatureCypherException("unsupported message type" + fromJsonToSignedMessageBean.getSignatureType());
-            }
-            if (verifyResult.isValid()) {
-                return returnObj;
-            } else {
-                throw new InvalidChatMessageSignatureException("invalid message signature");
-            }
-
-        } catch (JsonProcessingException ex) {
-            throw new ChatMessageException(ex);
-        }
-
-    }
 
 
     public static final RegisterUserRequestBean getSignedRegisteredUserRequests(InstanceWalletKeystoreInterface iwk, int i, RegisterUserRequestSignedContentBean registerUserRequestSignedContentBean) throws MessageException {
